+++ PraksaWebApp/Controllers/PatientsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using PraksaWebApp.Models;
 using System.Net.Http.Json;
 
@@ -6,6 +7,7 @@ using Microsoft.Extensions.Options;
 
 namespace PraksaWebApp.Controllers
 {
+    [Authorize]
     public class PatientsController : Controller
     {
         private readonly HttpClient _httpClient;
@@ -23,6 +25,13 @@ namespace PraksaWebApp.Controllers
         // Ги прикажува сите пациенти
         public async Task<IActionResult> Index(int page = 1)
         {
+            var tipClaim = User.FindFirst("tip_korisnik")?.Value;
+
+            if (tipClaim != "0")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (HttpContext.Session.GetString("username") == null)
             {
                 return RedirectToAction("Index", "Korisnik");
@@ -33,24 +42,19 @@ namespace PraksaWebApp.Controllers
                     _apiSettings.BaseUrl + "Patients"
                 );
 
-
             int pageSize = 10;
-
 
             int totalPages = (int)Math.Ceiling(
                 pacienti.Count / (double)pageSize
             );
-
 
             var pacientiNaStrana = pacienti
                 .Skip((page - 1) * pageSize)
                 .Take(pageSize)
                 .ToList();
 
-
             ViewBag.CurrentPage = page;
             ViewBag.TotalPages = totalPages;
-
 
             return View(pacientiNaStrana);
         }
@@ -60,23 +64,30 @@ namespace PraksaWebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Save(Patient patient)
         {
+            var tipClaim = User.FindFirst("tip_korisnik")?.Value;
+
+            if (tipClaim != "0")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             HttpResponseMessage response;
 
             if (patient.id == 0)
             {
                 // Нов пациент
                 response = await _httpClient.PostAsJsonAsync(
-                             _apiSettings.BaseUrl + "Patients",
-                               patient
-);
+                    _apiSettings.BaseUrl + "Patients",
+                    patient
+                );
             }
             else
             {
                 // Измена на постоечки пациент
                 response = await _httpClient.PutAsJsonAsync(
-                 $"{_apiSettings.BaseUrl}Patients?id={patient.id}",
-                     patient
- );
+                    $"{_apiSettings.BaseUrl}Patients?id={patient.id}",
+                    patient
+                );
             }
 
             var result = await response.Content.ReadAsStringAsync();
@@ -93,9 +104,16 @@ namespace PraksaWebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
+            var tipClaim = User.FindFirst("tip_korisnik")?.Value;
+
+            if (tipClaim != "0")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             await _httpClient.DeleteAsync(
-            $"{_apiSettings.BaseUrl}Patients?id={id}"
-                );
+                $"{_apiSettings.BaseUrl}Patients?id={id}"
+            );
 
             return RedirectToAction("Index");
         }
