+++ PraksaWebApp/Controllers/StatusController.cs
@@ -1,10 +1,12 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using System.Net.Http.Json;
 using PraksaWebApp.Models;
 
 namespace PraksaWebApp.Controllers
 {
+    [Authorize]
     public class StatusController : Controller
     {
         private readonly HttpClient _httpClient;
@@ -18,25 +20,38 @@ namespace PraksaWebApp.Controllers
             _apiSettings = apiSettings.Value;
         }
 
-
         public async Task<IActionResult> Index()
         {
+            var tipClaim = User.FindFirst("tip_korisnik")?.Value;
+
+            if (tipClaim != "0")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (HttpContext.Session.GetString("username") == null)
             {
                 return RedirectToAction("Index", "Korisnik");
             }
 
-            List<Status>? status = await _httpClient.GetFromJsonAsync<List<Status>>(
-                _apiSettings.BaseUrl + "Status"
-            );
+            List<Status>? status =
+                await _httpClient.GetFromJsonAsync<List<Status>>(
+                    _apiSettings.BaseUrl + "Status"
+                );
 
             return View(status);
         }
 
-
         [HttpPost]
         public async Task<IActionResult> Save(Status status)
         {
+            var tipClaim = User.FindFirst("tip_korisnik")?.Value;
+
+            if (tipClaim != "0")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             Console.WriteLine("ID: " + status.id);
             Console.WriteLine("NAME: " + status.status_name);
 
@@ -62,10 +77,16 @@ namespace PraksaWebApp.Controllers
             return RedirectToAction("Index");
         }
 
-
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
+            var tipClaim = User.FindFirst("tip_korisnik")?.Value;
+
+            if (tipClaim != "0")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             await _httpClient.DeleteAsync(
                 $"{_apiSettings.BaseUrl}Status?id={id}"
             );
